requirements.txt
fire
aim
pytorch-ignite
transformers
